+++ neurodamus/cell_distributor.py
@@ -28,7 +28,7 @@ from .metype import Cell_V5, Cell_V6, EmptyCell
 from .target_manager import TargetSpec
 from .utils import compat
 from .utils.logging import log_verbose, log_all
-from .utils.memory import get_mem_usage
+from .utils.memory import DryRunStats, get_mem_usage_kb
 
 
 class NodeFormat(Enum):
@@ -253,7 +253,7 @@ class CellManagerBase(_CellManager):
         return gidvec, me_infos, total_cells, full_size
 
     # -
-    def finalize(self, *_):
+    def finalize(self, **opts):
         """Instantiates cells and initializes the network in the simulator.
 
         Note: it should be called after all cell distributors have done load_nodes()
@@ -262,13 +262,13 @@ class CellManagerBase(_CellManager):
         if self._local_nodes is None:
             return
         logging.info("Finalizing cells... Gid offset: %d", self._local_nodes.offset)
-        self._instantiate_cells()
+        self._instantiate_cells(**opts)
         self._update_targets_local_gids()
         self._init_cell_network()
         self._local_nodes.clear_cell_info()
 
     @mpi_no_errors
-    def _instantiate_cells(self, _CellType=None):
+    def _instantiate_cells(self, _CellType=None, **_opts):
         CellType = _CellType or self.CellType
         assert CellType is not None, "Undefined CellType in Manager"
         Nd.execute("xopen_broadcast_ = 0")
@@ -286,49 +286,52 @@ class CellManagerBase(_CellManager):
             self._store_cell(gid + cell_offset, cell)
 
     @mpi_no_errors
-    def _instantiate_cells_dry(self, _CellType=None):
-        CellType = _CellType or self.CellType
+    def _instantiate_cells_dry(self, CellType, skip_metypes, **_opts):
+        """
+        Instantiates the subset of selected cells while measuring memory taken by each metype
+
+        Args:
+            CellType: The cell type class
+            full_memory_counter: The memory counter to be updated for each metype
+        """
         assert CellType is not None, "Undefined CellType in Manager"
         Nd.execute("xopen_broadcast_ = 0")
 
         logging.info(" > Dry run on cells... (%d in Rank 0)", len(self._local_nodes))
-        logging.info("Memory usage for metype combinations:")
         cell_offset = self._local_nodes.offset
-
         gid_info_items = self._local_nodes.items()
 
-        prev_emodel = None
-        prev_mtype = None
-        start_memory = get_mem_usage()
-        n_cells = 0
+        prev_metype = None
+        prev_memory = get_mem_usage_kb()
+        metype_n_cells = 0
         memory_dict = {}
+        MAX_CELLS = 50
+
+        def store_metype_stats(metype, n_cells):
+            nonlocal prev_memory
+            end_memory = get_mem_usage_kb()
+            memory_allocated = end_memory - prev_memory
+            log_all(logging.DEBUG, " * METype %s: %.1f KiB averaged over %d cells",
+                    metype, memory_allocated/n_cells, n_cells)
+            memory_dict[metype] = memory_allocated / n_cells
+            prev_memory = end_memory
 
         for gid, cell_info in gid_info_items:
-            diff_mtype = prev_mtype != cell_info.mtype
-            diff_emodel = prev_emodel != cell_info.emodel
-            first = prev_emodel is None and prev_mtype is None
-            if (diff_mtype or diff_emodel) and not first:
-                end_memory = get_mem_usage()
-                memory_allocated = end_memory - start_memory
-                log_all(logging.INFO, " * %s %s: %f MB averaged over %d cells",
-                        prev_emodel, prev_mtype, memory_allocated/n_cells, n_cells)
-                memory_dict[(prev_emodel, prev_mtype)] = memory_allocated/n_cells
-                start_memory = end_memory
-                n_cells = 0
-
+            metype = "{0.mtype}-{0.etype}".format(cell_info)
+            if metype in skip_metypes:
+                continue
+            if prev_metype is not None and metype != prev_metype:
+                store_metype_stats(prev_metype, metype_n_cells)
+                metype_n_cells = 0
+            if metype_n_cells >= MAX_CELLS:
+                continue
             cell = CellType(gid, cell_info, self._circuit_conf)
             self._store_cell(gid + cell_offset, cell)
+            prev_metype = metype
+            metype_n_cells += 1
 
-            prev_emodel = cell_info.emodel
-            prev_mtype = cell_info.mtype
-            n_cells += 1
-
-        if prev_emodel is not None and prev_mtype is not None:
-            end_memory = get_mem_usage()
-            memory_allocated = end_memory - start_memory
-            log_all(logging.INFO, " * %s %s: %f MB averaged over %d cells",
-                    prev_emodel, prev_mtype, memory_allocated/n_cells, n_cells)
-            memory_dict[(prev_emodel, prev_mtype)] = memory_allocated/n_cells
+        if prev_metype is not None and metype_n_cells > 0:
+            store_metype_stats(prev_metype, metype_n_cells)
 
         return memory_dict
 
@@ -539,12 +542,11 @@ class CellDistributor(CellManagerBase):
     def load_nodes(self, load_balancer=None, **kw):
         """gets gids from target, splits and returns a GidSet with all metadata
         """
-        loader_opts = kw.pop("loader_opts", {})
+        loader_opts = kw.pop("loader_opts", {}).copy()
         all_cell_requirements = SimConfig.cell_requirements
         cell_requirements = all_cell_requirements.get(self._population_name) or (
             self.is_default and all_cell_requirements.get(None)
         )
-
         if self._node_format == NodeFormat.SONATA:
             loader = cell_readers.load_sonata
             loader_opts["node_population"] = self._population_name  # mandatory in Sonata
@@ -555,24 +557,35 @@ class CellDistributor(CellManagerBase):
                 raise Exception('Additional cell properties only available with SONATA')
             nodes_filename = self._circuit_conf.CellLibraryFile
             loader = self._cell_loaders.get(nodes_filename, cell_readers.load_mvd3)
+            loader_opts = {}
 
         log_verbose("Nodes Format: %s, Loader: %s", self._node_format, loader.__name__)
         return super().load_nodes(load_balancer, _loader=loader, loader_opts=loader_opts)
 
-    def _instantiate_cells(self, *_):
+    def _instantiate_cells(self, dry_run_stats_obj: DryRunStats = None, **opts):
+        """
+        Instantiates cells, honouring dry_run if provided
+
+        Args:
+            dry_run_store_stats: Do a dry-run and update the inner fields accordingly
+        """
         if self.CellType is not NotImplemented:
             return super()._instantiate_cells(self.CellType)
         conf = self._circuit_conf
         CellType = Cell_V5 if self._is_v5_circuit else Cell_V6
         if conf.MorphologyType:
             CellType.morpho_extension = conf.MorphologyType
+
         log_verbose("Loading metypes from: %s", conf.METypePath)
         log_verbose("Loading '%s' morphologies from: %s",
                     CellType.morpho_extension, conf.MorphologyPath)
-        if SimConfig.dry_run:
-            super()._instantiate_cells_dry(CellType)
+        if dry_run_stats_obj is None:
+            super()._instantiate_cells(CellType, **opts)
         else:
-            super()._instantiate_cells(CellType)
+            cur_metypes_mem = dry_run_stats_obj.metype_memory
+            memory_dict = self._instantiate_cells_dry(CellType, cur_metypes_mem, **opts)
+            log_verbose("Updating global dry-run memory counters with %d items", len(memory_dict))
+            cur_metypes_mem.update(memory_dict)
 
 
 class LoadBalance: